requirements.txt
click==8.1.3
colorama==0.4.5
DateTime==4.7
Flask==2.2.2
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysql-connector-python-rf==2.2.2
mysqlclient==2.1.1
protobuf==3.20.1
PyMySQL==1.0.2
pytz==2022.2.1
PyYAML==6.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
zope.interface==5.4.0
